% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{reach_pal}
\alias{reach_pal}
\title{Return function to interpolate a drsimonj color palette}
\usage{
reach_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a drsimonj color palette
}
