% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_color_reach}
\alias{scale_color_reach}
\title{scale_color_reach}
\usage{
scale_color_reach(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in reach_palettes}

\item{discrete}{Boolean indicating whether to use discrete or continuous scale (default= T)}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Adds a custom reach palette to color
}
